requirements.txt
# Core AI/ML
torch>=2.0.0
transformers>=4.35.0
vllm>=0.2.0
accele rate>=0.24.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
sentence-transformers>=2.2.2

# Vector database
faiss-cpu>=1.7.4

# API and serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
requests>=2.31.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.1.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

